requirements.txt
llama-cpp-python==0.2.90
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
pypdf==3.17.1
requests==2.31.0
